Move IDrinkShortDescription
