+++ src/DrinksUI.Web/DrinksUI.Web.csproj
@@ -9,5 +9,4 @@
     <ProjectReference Include="..\DrinksUI.Dtos\DrinksUI.Dtos.csproj" />
   </ItemGroup>
 
-
 </Project>
