+++ src/DrinksUI.Web/Pages/Index.razor
@@ -2,6 +2,7 @@
 
 @using DrinksUI.Dtos
 @using DrinksUI.Data.Services
+@using DrinksUI.Dtos.Interfaces
 
 @inject DrinkService DrinkService
